render face contours

+++ app/src/main/java/uk/me/jeremygreen/merging/model/Coordinate.kt
@@ -23,6 +23,6 @@ data class Coordinate(
 
     val faceId: Long,
 
-    val x: Int,
-    val y: Int
+    val x: Float,
+    val y: Float
 )
